requirements.txt
requests>=2.28.0
google-api-core>=1.31.5
google-auth>=2.22.0
google-generativeai==0.8.5
google-auth-httplib2
praw>=7.7.0
python-dotenv>=1.0.0
firecrawl>=0.2.0
aiohttp>=3.8.4
tavily-python>=0.2.3
openai>=1.0.0
firebase-admin>=6.2.0
email-validator>=2.1.0
httpx-oauth>=0.10.0
pyjwt>=2.6.0
supabase>=2.0.0
pydantic>=2.0.0
pydantic-extra-types>=2.0.0  # For EmailStr
PyYAML>=6.0
asyncio>=3.4.3
fastapi
uvicorn==0.22.0
watchfiles==0.19.0
websockets
bcrypt
python-frontmatter==1.1.0
pillow
python-multipart
python-slugify==8.0.4
google-genai